server-service/requirements.txt
loguru==0.5.3
google_auth_oauthlib==0.4.6
pydantic==1.8.2
asyncpg~=0.27.0
cryptography==35.0.0
envfileparser==0.0.8
google_api_python_client==2.31.0
protobuf==3.19.1

requests~=2.28.2